backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
sqlmodel==0.0.16
asyncpg==0.29.0
redis==5.0.4
rq==1.16.1
boto3==1.34.113
pydantic-settings==2.3.1
stripe==10.5.0
python-multipart==0.0.9


